Move competition data loader to helpers and update imports
