+++ data_processing.py
@@ -834,90 +834,6 @@ def preprocess_data(X_train, X_test, use_autoencoder=False, autoencoder_latent_d
     return X_train_preprocessed, X_test_preprocessed
 
 
-def load_competition_data(train_path, test_path, supp_paths=None, use_supplementary=True):
-    """
-    Load competition data including main training, test, and optional supplementary datasets
-    
-    Args:
-        train_path: Path to main training data
-        test_path: Path to test data
-        supp_paths: List of paths to supplementary datasets
-        use_supplementary: Whether to include supplementary datasets
-    
-    Returns:
-        Tuple of (train_df, test_df)
-    """
-    print("Loading training data...")
-    
-    # Load main training data
-    train_df = pd.read_csv(train_path)
-    # Add new_sim flag - True for main dataset
-    train_df['new_sim'] = True
-    print(f"Main training data shape: {train_df.shape}")
-    
-    if use_supplementary and supp_paths:
-        # Load and combine supplementary datasets
-        print("\nLoading supplementary datasets...")
-        all_train_dfs = [train_df]
-        
-        for supp_path in supp_paths:
-            try:
-                supp_df = pd.read_csv(supp_path)
-                
-                # Handle dataset1 with Tc_mean column
-                if 'dataset1.csv' in supp_path and 'TC_mean' in supp_df.columns:
-                    # Rename TC_mean to Tc
-                    supp_df = supp_df.rename(columns={'TC_mean': 'Tc'})
-                    print(f"Renamed TC_mean to Tc in dataset1")
-                
-                # Add id column if missing
-                if 'id' not in supp_df.columns:
-                    # Create synthetic ids for supplementary data
-                    supp_df['id'] = [f'supp_{supp_path.split("/")[-1].split(".")[0]}_{i}' 
-                                     for i in range(len(supp_df))]
-                
-                # Add new_sim flag - False for supplementary datasets
-                supp_df['new_sim'] = False
-                print(f"Loaded {supp_path}: {supp_df.shape}")
-                all_train_dfs.append(supp_df)
-            except Exception as e:
-                print(f"Could not load {supp_path}: {e}")
-        
-        # Combine all training data
-        train_df = pd.concat(all_train_dfs, ignore_index=True)
-        print(f"\nCombined training data shape: {train_df.shape}")
-    else:
-        print("\n** Using ONLY main training data (no supplementary datasets) **")
-    
-    # Load test data
-    print("\nLoading test data...")
-    test_df = pd.read_csv(test_path)
-    # Add new_sim flag - True for test dataset (as it's from the competition)
-    test_df['new_sim'] = True
-    print(f"Test data shape: {test_df.shape}")
-    
-    # Remove duplicates from training data
-    print("\nRemoving duplicates from training data...")
-    original_count = len(train_df)
-    
-    # Count non-null target values for each row
-    target_columns = ['Tg', 'FFV', 'Tc', 'Density', 'Rg']
-    train_df['target_count'] = train_df[target_columns].notna().sum(axis=1)
-    
-    # Sort by target_count (descending) and new_sim (True first) to prioritize rows with more data
-    train_df = train_df.sort_values(['target_count', 'new_sim'], ascending=[False, False])
-    
-    # Keep first occurrence of each SMILES (which has the most target values)
-    train_df = train_df.drop_duplicates(subset=['SMILES'], keep='first')
-    
-    # Remove the temporary target_count column
-    train_df = train_df.drop('target_count', axis=1)
-    
-    duplicate_count = original_count - len(train_df)
-    print(f"Removed {duplicate_count} duplicate rows")
-    print(f"Final training data shape: {train_df.shape}")
-    
-    return train_df, test_df
